requirements_pip.txt
Django=1.5.4
django-sitetree==0.9.5
facebook-sdk==0.4.0
pytz==2013b